Don't fire expire_events from lease beacon
