+++ changelog/+leasebeacondupevents.fixed.md
@@ -1,0 +1,1 @@
+Fixed the `vault_lease` beacon firing duplicate lease expiry events on the Salt event bus when `cache:expire_events` was enabled
